refactor: move change emits into common file
as long as defineEmits is specified in the component itself, the other functions can be abstracted
