+++ src/components/inputs/Dropdown.vue
@@ -29,7 +29,7 @@
 import { useComponentId } from '../utils/compid.js';
 import { computed } from 'vue';
 import FontAwesomeIcon from '../../icons.js';
-import { debounce } from 'dettle';
+import { useChangeEmits } from './common.js';
 
 const props = defineProps({
   labelPosition: {
@@ -58,21 +58,8 @@ const dropdownOptions = computed(() => {
   return outputOptions;
 });
 
-let delay;
-try {
-  delay = parseInt(props.delay) || 0;
-} catch {
-  delay = 0;
-}
 const emit = defineEmits(['change', 'update:modelValue']);
-function _emitChange(event) {
-  emit('change', event);
-}
-const emitChange = debounce(_emitChange, delay);
-function valueChanged(event) {
-  emit('update:modelValue', event.target.value);
-  emitChange(event);
-}
+const { valueChanged } = useChangeEmits(emit, props.delay);
 </script>
 
 <style module lang="scss">
